Add ability to favorite entries with a star icon
Users can mark their favorite days with a star. When the star is clicked, it will fill in and be saved as "starred" in the database, indicating that it's been favorited. If the star is clicked again, the star icon will become hollow/empty and the database will be updated.